refactor to remove generateInstanceid
