+++ src/extensions/rolesHelper.ts
@@ -3,7 +3,7 @@ import * as api from "../api";
 import * as utils from "../utils";
 import { Role } from "./extensionsApi";
 import { iam } from "../gcp";
-import { getRandomString } from "./generateInstanceId";
+import { getRandomString } from "./utils";
 
 const API_VERSION = "v1";
